+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Fixed navigation bar (in anticipation of upcoming group/set functionality)
 - Scroll new or focused cards to the top of the view to give more space to the selectable options
+- New button icons
 
 ### Changed
 - New, more flexible data structure